updating basis function api
