+++ python/lsst/sims/featureScheduler/__init__.py
@@ -1,9 +1,2 @@
 from .core_scheduler import *
-from .features import *
-from .basis_functions import *
-from .surveys import *
-from .utils import *
-from .dithering import *
-from .comcamTessellate import *
-from .tsp import *
 from .version import *
