+++ python/lsst/sims/featureScheduler/core_scheduler.py
@@ -3,22 +3,24 @@ from builtins import object
 import numpy as np
 import healpy as hp
 from lsst.sims.utils import _hpid2RaDec
-from .utils import hp_in_lsst_fov, set_default_nside, hp_in_comcam_fov
+from lsst.sims.featureScheduler.utils import hp_in_lsst_fov, set_default_nside, hp_in_comcam_fov
+from lsst.sims.featureScheduler.features import Conditions
 import warnings
 import logging
 
-default_nside = None
+
+__all__ = ['Core_scheduler']
 
 
 class Core_scheduler(object):
     """Core scheduler that takes completed obsrevations and observatory status and requests observations.
     """
 
-    def __init__(self, surveys, nside=default_nside, camera='LSST'):
+    def __init__(self, surveys, nside=None, camera='LSST', conditions=None):
         """
         Parameters
         ----------
-        surveys : list of survey objects
+        surveys : list of lsst.sims.featureScheduler.survey objects
             A list of surveys to consider. If multiple surveys retrurn the same highest
             reward value, the survey at the earliest position in the list will be selected.
             Can also be a list of lists to make heirarchical priorities.
@@ -27,9 +29,17 @@ class Core_scheduler(object):
         camera : str ('LSST')
             Which camera to use for computing overlapping HEALpixels for an observation.
             Can be 'LSST' or 'comcam'
+        conditions : a lsst.sims.featureScheduler.features.Conditions object (None)
+            An object that hold the current conditions and derived values (e.g., 5-sigma depth). Will
+            generate a default if set to None.
         """
         if nside is None:
             nside = set_default_nside()
+
+        if conditions is None:
+            self.conditions = Conditions(nside=nside)
+        else:
+            self.conditions = conditions
 
         self.log = logging.getLogger("Core_scheduler")
         # initialize a queue of observations to request
@@ -80,7 +90,7 @@ class Core_scheduler(object):
             for survey in surveys:
                 survey.add_observation(observation, indx=indx)
 
-    def update_conditions(self, conditions):
+    def update_conditions(self, conditions_in):
         """
         Parameters
         ----------
@@ -88,11 +98,11 @@ class Core_scheduler(object):
             The current conditions of the telescope (pointing position, loaded filters, cloud-mask, etc)
         """
         # Add the current queue and scheduled queue to the conditions
-        conditions['queue'] = self.queue
-        for surveys in self.survey_lists:
-            for survey in surveys:
-                survey.update_conditions(conditions)
-        self.conditions = conditions
+        self.conditions.queue = self.queue
+
+        # Update the conditions
+        for key in conditions_in:
+            setattr(self.conditions, key, conditions_in[key])
 
     def request_observation(self):
         """
@@ -165,10 +175,11 @@ class Core_scheduler(object):
         self.is_sequence = False
         self.survey_index = [0, 0]
 
+
 class Core_scheduler_parallel(Core_scheduler):
     """Execute survey methods in parallel
     """
-    def __init__(self, surveys, nside=default_nside, camera='LSST'):
+    def __init__(self, surveys, nside=None, camera='LSST'):
         """
         Before running, start ipyparallel engines at the command line with something like:
         > ipcluster start -n 7
