+++ python/lsst/sims/featureScheduler/features/README.md
@@ -1,0 +1,4 @@
+
+XXXX
+
+Document the atributes that are in the conditions object.  They should be the same name as the conditions coming in from driver.py
